move runInTransaction tests
